fix(tui): 修复 TUI 代码审阅发现的 25 项问题
根据 docs/tui-design-specification.md 规范审阅结果，修复以下问题：

Critical:
- C1: EventKindFor 添加 ApprovalPromptChatEvent 映射
- C2: JSON 主题解析添加 system/assistantText/loaderSpinner/progressBar 4 个缺失 token

High:
- H1: 消除 itoa 三处重复，提取共享 internal.ITOA
- H2: watch.go 仅在加载成功时更新 lastMtime，失败时记录错误
- H3: tui_render.go 硬编码 ANSI 字符串改用 terminal/ansi.go 构造器
- H4: Editor/Input 选中色从硬编码 256 色索引改为主题驱动 (SelectionBg)
- H6: Component.Render 补充并发安全文档

Medium:
- M1: 补充 6 个导出 Msg 类型的 Godoc 注释
- M3: 移除 eventLoop 中冗余的 FlushEsc 调用
- M4: 修复 panic 路径中 stdin 未关闭的 goroutine 泄漏
- M5: watch.go 加载失败时记录错误日志
- M6: 修复 macOS Light 模式下系统外观检测返回 AppearanceUnknown 的 bug
- M7: QuantizeTheme 标注为 no-op 并补充说明

Low:
- L1-L10: 消除 15 个 lint 告警（gofmt/gocritic/staticcheck/unconvert 等）
- TerminalSupportsKittyKeyboard 添加 Deprecated 标记
- 补充 syntax_langs.go 文件级 Godoc, macOS 拼写修正

配套文档：
- docs/tui-design-specification.md （新增 TUI 设计规范）
- docs/review/tui-code-review-2026-07-25.md （新增审阅报告）

+++ tui/overlay.go
@@ -51,11 +51,11 @@ func hexToCoreColor(hex string) core.Color {
 			c := h[i*2+j]
 			switch {
 			case c >= '0' && c <= '9':
-				v = v*16 + uint8(c-'0')
+				v = v*16 + (c - '0')
 			case c >= 'a' && c <= 'f':
-				v = v*16 + uint8(c-'a'+10)
+				v = v*16 + (c - 'a' + 10)
 			case c >= 'A' && c <= 'F':
-				v = v*16 + uint8(c-'A'+10)
+				v = v*16 + (c - 'A' + 10)
 			default:
 				return core.Palette(235) // non-hex char → fallback
 			}
